pyproject.toml
[tool.poetry]
name = "simple-rdp"
version = "0.1.0"
description = "A Python RDP client for automation - exposes screen capture and input transmission"
authors = []
readme = "README.md"
license = "MIT"
packages = [{include = "simple_rdp", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
pyspnego = "^0.12.0"
pyasn1 = "^0.6.2"
pillow = "^12.1.0"
asn1crypto = "^1.5.1"
python-dotenv = "^1.0.0"
rle-fast = "^0.1.0"  # Rust RLE decompression (required)

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
ruff = "^0.14.14"
pre-commit = "^4.5.1"
pytest-asyncio = "^1.3.0"
mypy = "^1.19.1"
rich = "^14.3.1"
python-docx = "^1.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.ruff.lint.isort]
force-single-line = true

[tool.ruff.format]
quote-style = "double"

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
mypy_path = "src"
exclude = ["tests/", "agents/"]
ignore_missing_imports = true
